extract logic to helper-function
