+++ client/src/app/components/report-section/report-section.component.ts
@@ -1,8 +1,8 @@
 import { Component, Input } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { QuestionInput, QuestionType } from 'src/app/types/question-input.type';
+import { getOptionChoiceAnswer } from 'src/app/pages/report/report-helpers';
+import { QuestionType } from 'src/app/types/question-input.type';
 import { QuestionSection } from 'src/app/types/question-section.type';
-import { TranslatableString } from 'src/app/types/translatable-string.type';
 import { environment } from 'src/environments/environment';
 
 @Component({
@@ -20,18 +20,9 @@ export class ReportSectionComponent {
 
   public isDebug = environment.debug;
 
+  public renderOptionChoiceAnswer = getOptionChoiceAnswer;
+
   constructor(private route: ActivatedRoute) {
     this.programId = this.route.snapshot.params.id;
-  }
-
-  public renderOptionChoiceAnswer(
-    question: QuestionInput,
-    answer: string | string[],
-  ): string | TranslatableString {
-    const chosenOption = question.optionChoices.find(
-      (option) => option.name === answer,
-    );
-
-    return chosenOption.label;
   }
 }
